Improved OfflineExporting.
