+++ ts/Extensions/OfflineExporting/OfflineExportingVendor.d.ts
@@ -1,0 +1,40 @@
+/* *
+ *
+ *  (c) 2010-2021 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type G from '../../Core/Globals';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+declare global {
+    interface Canvg {
+        fromString(
+            ctx: CanvasRenderingContext2D,
+            svg: string
+        ): Canvg;
+        start(): void;
+    }
+    interface CanvgNamespace {
+        Canvg: Canvg;
+    }
+    class jsPDF { // eslint-disable-line @typescript-eslint/class-name-casing
+        constructor (a: string, b: string, c: Array<number>);
+        output: Function;
+    }
+}
